repositories rework

+++ app/src/main/java/com/example/deathnote/presentation/ui/screen/main_screen/components/statistics_screen_ui/StatsM_1.kt
@@ -33,7 +33,7 @@ fun StatsM_1(
     student: Student,
     respectfulAbsences: Int = 0,
     absences: Int = 0,
-    absencePercent: Float = 0f
+    absencePercent: Int = 0
 ) {
 
     Row(
